Enhance publish.sh

+++ publish.sh
@@ -2,7 +2,7 @@
 set -e
 cd SeaORM
 npm run build
-rm -r ../docs/SeaORM
+[[ -d ../docs/SeaORM ]] && rm -r ../docs/SeaORM
 mv build ../docs/SeaORM
 cd ..
 git add docs
